+++ Editor/HeaderGUI/DrawHeaderGUI.cs
@@ -283,16 +283,16 @@ namespace SaintsField.Editor.HeaderGUI
                     .ToList();
             }
 
-            HashSet<ISearchable> removeSaintsEditor = new HashSet<ISearchable>();
+            HashSet<object> removeSaintsEditor = new HashSet<object>();
 
-            // Debug.Log($"SearchableSaintsEditors={SearchableSaintsEditors.Count}");
+            // Debug.Log($"SearchableSaintsEditors={SearchableSaintsEditors.Count}; {string.Join(", ", SearchableSaintsEditors.Select(each => each.target))}");
 
-            foreach (ISearchable searchableSaintsEditor in SearchableSaintsEditors)
+            foreach (ISearchable searchableSaintsEditor in SearchableSaintsEditors.Values)
             {
                 // Debug.Log(searchableSaintsEditor);
                 if (!(Object)searchableSaintsEditor)
                 {
-                    removeSaintsEditor.Add(searchableSaintsEditor);
+                    removeSaintsEditor.Add(searchableSaintsEditor.target);
                     continue;
                 }
 
@@ -313,24 +313,29 @@ namespace SaintsField.Editor.HeaderGUI
                 //     height = useRect.height - 2,
                 // };
 
-                GUIStyle btnStyle = CacheAndUtil.GetIconButtonStyle();
+                GUIContent content = new GUIContent(EditorGUIUtility.IconContent("d_Search Icon"))
+                {
+                    tooltip = "Search Fields",
+                };
 
-                if (GUI.Button(useRect, GUIContent.none, btnStyle))
+                using EditorGUI.ChangeCheckScope change = new EditorGUI.ChangeCheckScope();
+                GUI.Toggle(
+                    useRect,
+                    searchableSaintsEditor.IsSearchableOn(),
+                    content,
+                    CacheAndUtil.GetIconButtonStyle()
+                );
+                if (change.changed)
                 {
                     searchableSaintsEditor.OnHeaderButtonClick();
                 }
-
-                string richLabel = searchableSaintsEditor.GetRichLabel();
-                if (!CacheAndUtil.ParsedXmlCache.TryGetValue(richLabel,
-                        out IReadOnlyList<RichTextDrawer.RichTextChunk> chunks))
-                {
-                    CacheAndUtil.ParsedXmlCache[richLabel] = chunks = RichTextDrawer.ParseRichXml(richLabel, "", null, null, firstTarget).ToArray();
-                }
-
-                CacheAndUtil.GetCachedRichTextDrawer().DrawChunks(useRect, GUIContent.none, chunks);
             }
 
-            SearchableSaintsEditors.ExceptWith(removeSaintsEditor);
+            foreach (object toRemove in removeSaintsEditor)
+            {
+                SearchableSaintsEditors.Remove(toRemove);
+            }
+            // SearchableSaintsEditors.Remove(removeSaintsEditor);
 
             if (renderTargetInfos.Count == 0)
             {
@@ -792,11 +797,17 @@ namespace SaintsField.Editor.HeaderGUI
             HeaderButtonDrawer.Update();
         }
 
-        private static readonly HashSet<ISearchable> SearchableSaintsEditors = new HashSet<ISearchable>();
+        private static readonly Dictionary<object, ISearchable> SearchableSaintsEditors = new Dictionary<object, ISearchable>();
 
         public static void SaintsEditorEnqueueSearchable(SaintsEditor saintsEditor)
         {
-            SearchableSaintsEditors.Add(saintsEditor);
+            SearchableSaintsEditors[saintsEditor.target] = saintsEditor;
+            // if (!SearchableSaintsEditors.Contains(saintsEditor))
+            // {
+            //     SearchableSaintsEditors.Add(saintsEditor);
+            //     Debug.Log($"add {saintsEditor.target}; newValue={SearchableSaintsEditors.Count}");
+            // }
+            // SearchableSaintsEditors.Add(saintsEditor);
         }
     }
 }